+++ DataCoreTests/GlobalUsings.cs
@@ -1,0 +1,12 @@
+// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+// Global using directives
+
+global using System.Collections.Generic;
+global using System.Linq;
+global using DataCore.Sql.Models;
+global using DataCore.Sql.TableScaleModels;
+global using FluentValidation.Results;
+global using NSubstitute;
+global using NUnit.Framework;
+global using static DataCore.ShareEnums;