+++ DataCoreTests/Sql/DataModels/BarcodeHelperTests.cs
@@ -2,9 +2,7 @@
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
 using DataCore.Sql.DataModels;
-using NUnit.Framework;
 using System;
-using NSubstitute;
 
 namespace DataCoreTests.Sql.DataModels;