+++ DataCoreTests/Sql/Models/CrudControllerTests.cs
@@ -1,10 +1,6 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-using System.Collections.Generic;
-using DataCore.Sql.TableScaleModels;
-using NUnit.Framework;
-
 namespace DataCoreTests.Sql.Models;
 
 [TestFixture]