+++ DataCoreTests/Sql/TableScaleModels/AccessValidatorTests.cs
@@ -1,13 +1,7 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-using DataCore.Sql.TableScaleModels;
-using FluentValidation.Results;
-using NSubstitute;
-using NUnit.Framework;
 using System;
-using System.Linq;
-using static DataCore.ShareEnums;
 
 namespace DataCoreTests.Sql.TableScaleModels;