+++ DataCoreTests/Sql/TableScaleModels/ProductionFacilityEntityTests.cs
@@ -1,17 +1,6 @@
 ﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-using DataCore.Sql.Models;
-using DataCore.Sql.TableScaleModels;
-using NUnit.Framework;
-using System.Collections.Generic;
-using System.Linq;
-using NSubstitute;
-using static DataCore.ShareEnums;
-// ReSharper disable MethodTooLong
-// ReSharper disable CognitiveComplexity
-// ReSharper disable ExcessiveIndentation
-
 namespace DataCoreTests.Sql.TableScaleModels;
 
 [TestFixture]
@@ -34,43 +23,6 @@ internal class ProductionFacilityEntityTests
             success = item.EqualsDefault();
             // Assert.
             Assert.True(success);
-        });
-    }
-
-    [Test]
-    public void Entity_Crud_DoesNotThrow()
-    {
-        Assert.DoesNotThrow(() =>
-        {
-            foreach (bool isShowMarkedItems in TestsEnums.GetBool())
-            {
-                List<BaseEntity>? items = TestsUtils.DataAccess.Crud.GetEntities<ProductionFacilityEntity>(
-                        (isShowMarkedItems == true) ? null
-                            : new FieldListEntity(new() { new(DbField.IsMarked, DbComparer.Equal, false) }),
-                        new(DbField.User, DbOrderDirection.Asc),
-                        10)
-                    ?.ToList<BaseEntity>();
-                if (items != null)
-                {
-                    List<ProductionFacilityEntity> itemsCast = items.Select(x => (ProductionFacilityEntity)x).ToList();
-                    if (itemsCast.Count > 0)
-                    {
-                        foreach (ProductionFacilityEntity item in itemsCast)
-                        {
-                            ProductionFacilityEntity itemCopy = item.CloneCast();
-                            Assert.AreEqual(true, item.Equals(itemCopy));
-                            Assert.AreEqual(true, itemCopy.Equals(item));
-                            ProductionFacilityEntity itemChange = new()
-                            {
-                                IsMarked = true,
-                            };
-                            _ = itemChange.ToString();
-                            Assert.AreNotEqual(true, itemChange.Equals(item));
-                            Assert.AreNotEqual(true, item.Equals(itemChange));
-                        }
-                    }
-                }
-            }
         });
     }
 }